+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v2.6.7] - 2026-09-23
+
+- ADD: Template default-images-manual-update
+- UPDATE: sto-sast-register-repo template IDP Workflow to fix library path issues
+
 ## [v2.6.6] - 2026-09-15
 
 - UPDATE: All templates to remove dependency on `hashicorp.time_sleep` in favor of `harness.harness_time_sleep`